create-a-basic-telegram-bot-using-2/requirements.txt
aiogram==3.27.0
python-dotenv==1.1.1
SQLAlchemy==2.0.49
APScheduler==3.11.2
